Rename using dash
https://github.com/Jiiks/BetterDiscordApp/issues/473
