+++ src/client/modules/plugin-manager.js
