+++ src/client/modules/theme-manager.js
